rewrite part 1
